requirements.txt
# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-vertexai>=2.0.0
langsmith>=0.1.0

# Google Vertex AI
google-cloud-aiplatform>=1.60.0
google-auth>=2.0.0

# Mail
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0  # Gmail API
msal>=1.0.0                      # Outlook / MS Graph

# Crawling
playwright>=1.40.0
httpx>=0.27.0
beautifulsoup4>=4.12.0

# Telegram
python-telegram-bot>=21.0.0

# Database
asyncpg>=0.29.0
redis>=5.0.0
pgvector>=0.2.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Scheduler
apscheduler>=3.10.0,<4.0.0

# Infra
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0                  # retry 로직
structlog>=24.0.0                # structured logging
